警告を全て対処した(概ね暗黙的型変換だった)

+++ CCostumeFactory.h
@@ -17,7 +17,7 @@ class CCostumeFactory {
   CCostumeFactory();
   CCostumeBase* create(std::string);
   CCostumeBase* create(int);
-  int getSize() { return list.size(); }
+  int getSize() { return (int)list.size(); }
   void getMinMaxFriction(double& min, double& max) {
     min = 9999;
     max = 0;
