+++ CDrawCircleGauge.cpp
@@ -2,11 +2,11 @@
 
 #include <DxLib.h>
 
-CDrawCircleGaugeReserve::CDrawCircleGaugeReserve(int GHandle, double priority,
+CDrawCircleGaugeReserve::CDrawCircleGaugeReserve(int GHandle, int priority,
                                                  int x, int y, double ratio,
                                                  int width, int height)
     : IRenderReserve(GHandle, priority, x, y, width, height), ratio(ratio) {}
 
 void CDrawCircleGaugeReserve::Render() const {
-  DrawCircleGauge(x, y, ratio * 100, GHandle);
+  DrawCircleGauge((int)x, (int)y, ratio * 100, GHandle);
 }
