+++ CDrawExtendWithBlend.h
@@ -2,12 +2,12 @@
 #include "IRenderReserve.h"
 class CDrawExtendWithBlendReserve : public IRenderReserve {
  protected:
-  float x1, y1, x2, y2;
+  double x1, y1, x2, y2;
   int Color, Blendmode, value;
 
  public:
-  CDrawExtendWithBlendReserve(int GHandle, double Priority, float x1, float y1,
-                              float x2, float y2, int color, int Blendmode,
+  CDrawExtendWithBlendReserve(int GHandle, int Priority, double x1, double y1,
+                              double x2, double y2, int color, int Blendmode,
                               int value);
   void Render() const;
 };
