+++ CDrawGraphFReserve.cpp
@@ -2,8 +2,8 @@
 
 #include <DxLib.h>
 
-CDrawGraphFReserve::CDrawGraphFReserve(int GHandle, float x, float y,
-                                       double Priority, int width, int height)
+CDrawGraphFReserve::CDrawGraphFReserve(int GHandle, double x, double y,
+                                       int Priority, int width, int height)
     : IRenderReserve(GHandle, Priority, x, y, width, height) {}
 
-void CDrawGraphFReserve::Render() const { DrawGraphF(x, y, GHandle, TRUE); }
+void CDrawGraphFReserve::Render() const { DrawGraphF((float)x, (float)y, GHandle, TRUE); }
