+++ CDrawGraphFReserve.h
@@ -2,7 +2,7 @@
 #include "IRenderReserve.h"
 class CDrawGraphFReserve : public IRenderReserve {
  public:
-  CDrawGraphFReserve(int GHandle, float x, float y, double Priority, int width,
+  CDrawGraphFReserve(int GHandle, double x, double y, int Priority, int width,
                      int height);
   void Render() const;
 };
