+++ CDrawRotaGraphReserve.cpp
@@ -4,7 +4,7 @@
 
 #include <cmath>
 
-CDrawRotaGraphReserve::CDrawRotaGraphReserve(int GHandle, double Priority,
+CDrawRotaGraphReserve::CDrawRotaGraphReserve(int GHandle, int Priority,
                                              int x, int y, float angle,
                                              float extend, int width,
                                              int height)
@@ -16,5 +16,5 @@ CDrawRotaGraphReserve::CDrawRotaGraphReserve(int GHandle, double Priority,
       extend(extend) {}
 
 void CDrawRotaGraphReserve::Render() const {
-  DrawRotaGraph(x, y, extend, angle, GHandle, TRUE);
+  DrawRotaGraph((int)x, (int)y, extend, angle, GHandle, TRUE);
 }
