+++ CEffect_MoneyNumber.cpp
@@ -15,17 +15,17 @@ void CEffect_MoneyNumber::Update() {
 }
 
 void CEffect_MoneyNumber::Render() const {
-  int l = std::log10(num) + 1;
+  int l = (int)std::log10(num) + 1;
   int n = num;
-  int x = Position.x + l * 5;
+  int x = (int)Position.x + l * 5;
   for (int i = 0; i < l; i++) {
     CImageManager::getIns()
         .find("effect_number3")
-        ->DrawRota(x, Position.y, 0.0, 1.0, Constant::priority_number, n % 10);
+        ->DrawRota(x, (int)Position.y, 0.0, 1.0, Constant::priority_number, n % 10);
     x -= 10;
     n /= 10;
   }
   CImageManager::getIns()
       .find("effect_number3")
-      ->DrawRota(x, Position.y, 0.0, 1.0, Constant::priority_number, 10);
+      ->DrawRota(x, (int)Position.y, 0.0, 1.0, Constant::priority_number, 10);
 }
