+++ CSound.cpp
@@ -9,5 +9,5 @@ void CSound::Play(int PlayType) { PlaySoundMem(SHandle, PlayType); }
 void CSound::Stop() { StopSoundMem(SHandle); }
 
 void CSound::SetVolume(double volume) {
-  ChangeVolumeSoundMem(volume * 255, SHandle);
+  ChangeVolumeSoundMem(int(volume * 255), SHandle);
 }
